% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_8day_to_yearly.R
\name{aggregate_8day_to_monthly}
\alias{aggregate_8day_to_monthly}
\title{Aggregate 8-day raster data to monthly totals}
\usage{
aggregate_8day_to_monthly(x, agg_fun = sum, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A SpatRaster with 8-day time series data.}

\item{agg_fun}{Aggregation function, e.g., sum or mean.}

\item{na.rm}{Logical, whether to remove NA values.}

\item{...}{Additional arguments passed to terra::tapp.}
}
\value{
A SpatRaster aggregated to monthly scale.
}
\description{
Aggregate 8-day raster data to monthly totals
}
