% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_8day_to_yearly.R
\name{aggregate_8day_to_yearly}
\alias{aggregate_8day_to_yearly}
\title{Aggregating SparRaster from 8-day to yearly}
\usage{
aggregate_8day_to_yearly(x, agg_fun = sum, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a 8-day sacle SparRaster}

\item{agg_fun}{aggregate function}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}

\item{...}{additional arguments passed to \code{fun}}
}
\value{
SpatRaster
}
\description{
Aggregating SparRaster from 8-day to yearly
}
