% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ext_crs.R
\name{convert_ext_crs}
\alias{convert_ext_crs}
\title{Convert crs of extent}
\usage{
convert_ext_crs(xmin, xmax, ymin, ymax, crs = "ESRI:54030")
}
\arguments{
\item{xmin}{}

\item{ymax}{}
}
\description{
Convert crs of extent
}
