% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date.R
\name{filter_date}
\alias{filter_date}
\title{Filter SpatRaster layers by time attributes}
\usage{
filter_date(r, start = NULL, end = NULL)
}
\arguments{
\item{r}{a SpatRaster object}

\item{start}{start date of SpatRaster (20250918)}

\item{end}{end date of SpatRaster (20250919)}
}
\value{
SpatRaster
}
\description{
Filter SpatRaster layers by time attributes
}
