% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshade.R
\name{hillshade}
\alias{hillshade}
\title{calculate hillshade}
\usage{
hillshade(x, direction = c(285, 15, 60, 330))
}
\arguments{
\item{x}{SpatRaster, single layer with elevation values. Values should have the same unit as the map units, or in meters when the crs is longitude/latitude}
}
\value{
Description of return value
}
\description{
calculate hillshade
}
