% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_multifiles.R
\name{is_dir}
\alias{is_dir}
\title{Check if the path is a folder}
\usage{
is_dir(path)
}
\arguments{
\item{path}{file or folder path (character)}
}
\value{
logical
}
\description{
Check if the path is a folder
}
