% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rast_list.R
\name{merge_rast_list}
\alias{merge_rast_list}
\title{Merging multiple SpatRasters}
\usage{
merge_rast_list(l)
}
\arguments{
\item{l}{A list containing multiple SpatRaster}
}
\value{
SpatRaster
}
\description{
Merging multiple SpatRasters
}
