% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_convert_tools.R
\name{rast2df}
\alias{rast2df}
\title{Converting SpatRaster to data.table}
\usage{
rast2df(r, melt = FALSE, ...)
}
\arguments{
\item{r}{SpatRaster}

\item{melt}{whether or not to use the 'melt' function on data.table}

\item{...}{any other arguments to be passed to/from other methods.}
}
\value{
data.table
}
\description{
Converting SpatRaster to data.table
}
