% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast2significance.R
\name{rast2significance}
\alias{rast2significance}
\title{Generating a function that can convert rast to significane hatch line}
\usage{
rast2significance()
}
\value{
function
}
\description{
Generating a function that can convert rast to significane hatch line
}
