% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_multifiles.R
\name{rast_multifiles}
\alias{rast_multifiles}
\title{Reading multiple files with terra::rast}
\usage{
rast_multifiles(path, merge = TRUE, ...)
}
\arguments{
\item{path}{multiple file paths or folder path (character)}

\item{merge}{whether to merge into one SpatRaster}

\item{...}{additional arguments passed on to the \code{rast,missing-method}}
}
\value{
a list or SpatRaster
}
\description{
Reading multiple files with terra::rast
}
